Feature/initial

+++ internal/role/controller.go
@@ -59,6 +59,9 @@ func (ctr *ControllerImpl) Get(c *gin.Context) error {
 // GetAssignedRole Returns assigned role for auth_id in path
 func (ctr *ControllerImpl) GetAssignedRole(c *gin.Context) error {
 	id := c.Param("auth_id")
+	if id == "" {
+		return apierror.NewBadRequestApiError("Empty auth ID")
+	}
 
 	role, err := ctr.service.GetAssignedRole(id)
 	if err != nil {
@@ -75,7 +78,7 @@ func (ctr *ControllerImpl) AssignRole(c *gin.Context) error {
 	if err := c.BindJSON(&body); err != nil {
 		return apierror.NewBadRequestApiError("Error reading body")
 	}
-	if body.AuthID == "" || body.RoleID < 0 {
+	if body.AuthID == "" || body.RoleID <= 0 {
 		return apierror.NewBadRequestApiError("Invalid auth ID or role ID")
 	}
 
@@ -93,6 +96,9 @@ func (ctr *ControllerImpl) AssignRole(c *gin.Context) error {
 // DeleteAssignedRole Deletes all assigned roles for given auth ID
 func (ctr *ControllerImpl) DeleteAssignedRole(c *gin.Context) error {
 	id := c.Param("auth_id")
+	if id == "" {
+		return apierror.NewBadRequestApiError("Empty auth ID")
+	}
 
 	err := ctr.service.DeleteAssignedRole(id)
 	if err != nil {
